Only generate TOC at top of singlefile doc HTML
Generating a bunch of TOCs at the Part level per DocBook default is
kind of disorienting when reading the single-file HTML output.  The
new Titan style has a "jump-to-top-of-page" in the lower-right-hand
corner of the viewport at all times anyway, and that or the Home key
are probably handier for TOC navigation than inserting a bunch of
miniTOCs inline anyway.

TOC generation in the chunked output is unchanged.