Adding float:right TOC in doc chapter pages
This only affects chunked (one-page-per-chapter) HTML doc output, not
the single-page documentation.

This commit introduces the necessary styling tweaks for putting
github-wiki-like floating TOCs on chapter pages, including overriding
section headings' inline clear:both style with a !important CSS rule,
but the content is not particularly useful.  It's just a TOC of the
chapter page, rather than a TOC of surrounding chapters/parts.