BerkeleyJEBlueprintsTest fixes
Modified cleanUp(), pre-suite, and post-suite methods to detect and
warn on graph leaks in the underlying Blueprints test methods.  Also
made cleanUp() and the post-suite method stop leaking unclosed store
managers and made cleanUp() stop behaving like a no-op for the first
test in the suite (important for repeated runs without a `mvn clean`
in between).