Fix regression in LogTest$MessagePuller
The most important change in this commit fixes a regression that I
introduced yesterday via 7da161f9d8f2ef9f3384abcc97794df63c00412b.
Specifically, it broke the calculation of the end of MessagePuller's
column slice bounds by allowing them to straddle two timeslices.
Timeslices are mapped 1-to-1 with keys, and each individual
MessagePuller run is only supposed to query one timeslice/key.  The
regression manifested as dropping messages in the broken column slice
belonging that belonged to the later of the two timeslices.

This commit also introduces a bunch of new logging output and a
toString() implementation that I used when investigating the
regression and which will probably be useful in future debugging.