Renaming a class
Prior to this commit, titan05 had two classes named
WriteConfiguration.  One belonged to titan-core's internal graph
ConfigNamespace/ConfigOption system, whereas the other was a class
with a main method that recursively walked a configuration object and
printed its members in AsciiDoc table format.

This commit renames the latter to ConfigurationPrinter in an attempt
to avoid future confusion.