Create HBase table before reading rowkey partition
This avoids a nonfatal but scary-looking exception when Titan is
opened against a HBase cluster for the first time.  The
HBaseStoreManager didn't check that Titan's table existed when reading
the rowkey partitions for the table.  It would attempt to lookup
rowkey bounds for a nonexistent table, leading to a
TableNotFoundException.  It would proceed to create the table and this
wouldn't occur on subsequent opens.

This should be fixed, since getLocalKeyPartition() now ensures that
the table exists before trying to read its rowkey bounds.