Major deb/rpm packaging update
Titan's debian and rpm packages now install and run again under Ubuntu
14.04 and CentOS 6.5 respectively. Tested alongside the packages for
ES and Cassandra and they seem to play nicely (1.0.3 and 2.0.7 resp).
Haven't tried to make this play nicely with Hadoop packages from any
distributors, though.

I also tweaked the apt/yum repo generation script to make it a bit
more robust and allow updating one repo type at a time instead of
doing both unconditionally.

The packaging controlfiles/specfiles and attendant scripts are still
setup with `mvn -pl titan-dist -Ppkg-tools install` from the root of
the repo and executed with {debian,redhat}/build.sh or
pkgtools/bin/build-all.sh.

I also did away with subpackages for each of the storage and indexing
backends.  titan-cassandra, titan-hbase, titan-es, ... no longer exist
as separate rpm and deb files; there's just one giant titan package
with all of the backend jars. The maven configuration bits and shell
scripts for generating a partition of titan's dependency tree and a
concomitant set of constellation of little subpackages with
nonoverlapping jar sets will live on in the git history in case we
decide to restore it in the future.  I don't think we want to sign on
to maintain that level of complexity just yet though.