Fixup log4j configuration shipped in deb/rpm pkgs
Now /etc/titan/log4j.properties exists and is respected by Rexster.

This file sets up a rolling file appender with date-and-timestamps.
These stamps include the yyyy-mm-dd date and the time of day to the
milliseconds place.  The default Rexster log4j configuration does not
include any kind of time or datestamp and I think one is sorely needed
for the rpm/deb distribution.

Implementation note: the only way to get Rexster to use a log4j config
is to either put it in a file named "log4j.properties" in the current
working directory of the JVM or to pack it in rexster's own jar so
that it will load when Rexster calls
Application.class.getResource("log4j.properties").  This behavior is
hardcoded into a static initializer in Rexster's Application class and
it overrides the default log4j configuration mechanism.  I would
rather use the default log4j mechanism but changing Rexster's behavior
is out-of-scope.