Porting introduction section of Faunus manual
I've brought over most of the images as-is.  This commit is slightly
more than a pure markup conversion: it also includes a little bit of
editing to attempt to replace references to "Faunus" with
"Titan-Hadoop" or just "Titan" if the context already strongly
connotes Hadoop.  I also dropped or spliced some content that only
made sense for Faunus as an independent project (like its own
downloads page and separate javadoc URL).