Deleting some cassandrathift cruft
This class has retained a couple of methods used in testing Titan
against a cluster of multiple Cassandra instances on the same physical
machine long past the removal of that test configuration from the
Maven build.  The method implementations are so old and untested at
this point that I wouldn't readily trust them to do the right thing on
the current Cassandra API; one of the two methods axed in this
commitwas adapted from the Cassandra 0.7 codebase.  Better to
reimplement if we need that functionality again in the future.