Resolving tx conflicts under serializable isolation
This commit should fix an issue with StandardTitanGraph.commit that
caused transaction conflicts on storage backends with serializable
isolation (but not on backends with weaker isolation guarantees).  It
makes STG.commit only open a secondary transaction for schema
mutations when the backend has weak isolation; when the backend
provides strong isolation, it sticks to a single tx now.

This commit also changes the isolation level used in most of the BDB
tests to serializable (from repeatable read, the default).

This commit still at least two known problems:

* TransactionalTitanGraphTestSuite.testCompetingThreads fails
  spectacularly with BDB configured for serializable isolation

* BerkeleyPartitionGraphTest.testVertexPartitioning
  nondeterministically fails.  These pseudorandom failures predate
  this commit; they are present on the ancestor commit also.  On my
  machine, 33 of 60 successive runs passed while the remaining 27
  failed with a few different assertions, all of which seem related to
  bogus values coming out of the fulgora OLAP result object.