Improve edge case in HBase's ensureTableExists
This method assumed that the table was enabled.  If the table was
already disabled for some reason -- say, because Titan was killed
before it could completely setup schema the first time it was executed
-- then this method would fail unnecessarily.  Now the method checks
whether the table is already enabled.  Besides this check, if the
method catches a TableNotEnabledException when disabling, it now logs
the and proceeds as usual instead of propagating it up the stack.