Refatoring titan-hadoop into -hadoop1 and -hadoop2
Faunus/Titan-Hadoop is now two separate artiacts with different
classifiers.  The user's expected to choose the appropriate
classifier.

This commit retains the reflective instantiation chicanery from my
previous commit's all-in-one attempt.  That could be simplified away
in future commits.

One particular test, BlueprintsGraphOutputMapReduceTest, is completely
broken in this commit and I'm still not sure why.  It's got a
nonsensical Hadoop filesystem configuration issue and a scary-looking
type error involving a Mockito synthetic object.  Something is
seriously wrong with that test (and potentially indicative of a wider
problem), but I'm not sure what.