Fix bug in ExpectedValueCheckingTx
EVCT calls getSlice with bounds [col, nextBiggerBuffer(col)), where
col is given by the caller.  This slice returns exact matches of col,
if present in the backend, but it can also return longer columns for
which col is a prefix.  This is a bug -- EVCT should be responsible
for ignoring columns which aren't an exact match of col provided by
the caller.

This manifested as a bug in
ThriftGraphCacheTest#testGraphConfiguration which actually had nothing
to do with db-cache.  The problem was that ECVT was considering both
values for "cache.db-cache" and "cache.db-cache-time" when attempting
to lock and check "cache.db-cache", causing a nonsensical
PermanentLockingException.