Refactor EVCS lock-checking logic
This change is mostly a refactoring and javadoc commit in the
ExpectedValueChecking* classes.

However, it also includes a functional change.  Previous to this
commit, locks taken on stores without any mutations (stores treated as
read-only) would be ignored.  Now locks are checked regardless of
whether the underlying store has any data mutations waiting.