Fix titan-hadoop-1's hbase deps
titan-hadoop-1 erroneously pulled in hbase 0.98.3-hadoop2 artifacts
with scope provided from titan-hadoop-parent.  This in turn pulled in
a bunch of Hadoop 2.2.0 jars.  I've changed titan-hadoop-1 to use
HBase 0.94.12 and left titan-hadoop-2 and -core on 0.98.3-hadoop2 for
the time being.

It would be nice to pull out the 0.98.3-hadoop2 and 0.94.12 version
constants as maven pom properties in a future commit, so that they
could be easily changed.