Fix configuration for Hadoop 2 combiners
This commit fixes a failsafe integration test failure in
TitanCassandraOutputFormatTest#testBulkEdgeDerivations that occurred
on titan-hadoop-2 but not on titan-hadoop-1.  The link step was
failing with an NPE when attempting to read the traversal direction
from the config under H2 (but succeeding under H1).  I think this is
due to the way Hadoop2Compiler handles job configuration; it sets
custom config keys, including the traversal direction for link steps,
on the reducer and mapper but not on the combiner.  This commit tries
to pull the custom config keys up to the job level so that they're
inherited by the combiner as well as the reducer and mapper.  At
least, I think that's what's going on -- not 100% confident I've got
the analysis correct, even though the fix seems to work.