Cleanup titan-hbase source/javadoc jar creation
Prior to this commit, titan-hbase's javadoc jar was built by copying
the source files for the titan-hbase-core and -09X modules into a
single directory, then invoking maven-javadoc-plugin on the directory.
This worked, but maven-javadoc-plugin would fill the console with
scary-looking though harmless warnings since it couldn't resolve any
of the imports in the source files.  The point of this commit is
basically to avoid those warnings.  The titan-hbase javadoc jar is now
built via an assembly that unpacks the titan-hbase-core javadoc.  This
is kind of a downgrade, since the javadoc no longer includes the
compat shim classes in the titan-hbase-09X modules, but there's not
much interesting content to javadoc in there anyway.  We could get
them back by creating yet another module, but we already have so
many...

The titan-hbase source jar is now generated by an assembly that
unpacks the source jars from titan-hbase-core and the -09X compat
modules.