Hack around Tokens static init error in testing
This commit should allow IDE execution of individual tests in
titan-hadoop-core that don't depend on a particular Hadoop version
(e.g. works with 2 but not 1) or try to run a MapReduce job.  This
could be done a little cleaner, but can't really test a fancier
solution until the compile errors are resolved anyway.