Add hack for millisecond sleepPast method
Mutation code in the storage backends assume that it's safe to play
with the least significant bit of timestamp longs, masking it to zero
for deletions, or'ing one for additions.  This is safe at microsecond
and nanosecond resolution, but milliseconds are long enough that
masking the lsb (1 ms) can result in write corruption that takes the
form of "lost" operations in testing.  This change modifies the
sleepPast method called by the mutation logic to compensate.