Use Titan schema provider in a output test method
This commit also renames several helper methods defined in
TitanOutputFormatTest.  These methods started with the string "test"
despite not actually constituting tests.  Also, these not-test methods
overload the names of actual test methods declared in subclasses.

Some part of these circumstances seemed to make the JUnit runner or
failsafe plugin or something seems to go nuts when I tried to run a
single one of these tests with -Dit.test=....  The runner would
execute the method overload defined in the subclass twice (not once in
the subclass overload and once in the super) and I don't really know
why.  Renaming these methods is a good idea anyway, if just to prevent
confusion about their purpose.  So now they start with "do" instead of
"test".