Faunus bug and test resource leak fixes
This commit changes FaunusElement.setResource to delete all existing
relations of the given type, even for multi-valued types.  This
differs from the old behavior of just dying with a precondition
failure on multi-valued types.  If the type is multi-valued and at
least one existing relation was deleted, then a warning is issued with
some context information and suggesting addRelation.

This commit also fixes several graph and one backend leak that caused
socket contention problems with Cassandra.