Adding test of EVCS transaction handling
EVCS = ExpectedValueCheckingStore and associated manager and
transaction classes.  These classes hold a transaction which is
supposed to provide global strong consistency and another transaction
which does not guarantee any particular level of consistency.  This
new test checks that the strong consistency tx is used when reading
data to check expectations and when writing mutations for a
transaction which has taken at least one lock.