Support external calls to titan-hbase test harness
The test support code in titan-hbase-core can start and stop HBase
daemons at several different supported versions.  Prior to this
commit, the harness assumed that the current working directory of the
process was one of the titan-hbase-parent/titan-hbase-09X/
directories.  Now the harness's assumption can be overridden by
setting the test.hbase.parentdir system property.  This supports reuse
of the HBase daemon test harness from other modules, such as
titan-dist.  It's not the cleanest way to do it, but it works for now.