Prevent avoidable NPE in TitanGraphOutputMR
TitanGraphOutputMapReduce has a static helper that constructs a
configuration with some default settings commonly used in testing.
These are Cassandra specific, which doesn't really belong in the
store-agnostic TitanGraphOutputMapReduce class anyway, but moving it
is a task for a future commit.  This commit just modifies the
default-setting logic to avoid an NPE when cassandra.yaml is not
available as a classloader resource in the same package as
TitanCassandraInputFormat.  That's still the first place it looks, but
if it comes up null, it tries the system classloader, and if that
comes up null, then it does not attempt to set a default value for the
cassandra.yaml path.