Fix and extend titan-dist tests
This commit should ideally be split up into smaller commits, but
overlapping changes in titan-dist/pom.xml made that difficult.

This commit fixes the existing Expect-based tests, all of which were
broken before this commit.  There were a couple of problems:

1. Vertex IDs assigned by Titan during even trivial gremlin.sh
examples have changed since the Expect tests were originally written,
so that the current vertex IDs disagree with the constants embedded in
the Expect scripts.  I replaced all of those constant vertex ID
expectations with regexes for nonnegative integers.

2. A mismatch between the config files generated in titan-dist/pom.xml
and the default config files used by CassandraStorageSetup prevented
Cassandra from even starting

Both of these are fixed.

This commit also adds three new tests of Faunus:

* graphson->cassandra, shutdown, then cassandra->graphson
* graphson->hbase, shutdown, then hbase->graphson
* graphson->graphson

The Cassandra tests run once on BOP and then again on Murmur.  The
HBase tests run against whatever HBase version is in the zipfile
(currently: 0.98).  The tests start HBase automatically using the test
harness in titan-hbase-parent.

This commit also disables batch-loading in the default
titan-hbase-output.properties and titan-cassandra-output.properties
files shipped in the zipfile.  This is a temporary workaround for the
schema inferencer not forcing batch-loading off.

This commit also adds hbase-server to the zipfile dependency tree.
The zipfile was originally written with 0.94, which just had a single
artifact named "hbase".  0.96 and 0.98 are split between the
"hbase-client" and "hbase-server" artifacts, and HBase's MapReduce
input format ships in the latter.  Failing to include it in the
zipfile produced ClassNotFoundException when attempting to use the
HBase input format.

Several sample Faunus config files included in the zipfile contain a
path to ./graph-of-the-gods.json.  I changed this to
example-data/graph-ofthe-gods.json, so that the example files work
without any modification or file copying when using MapReduce's local
(in-process) job runner and directly accessing the local filesystem.

The modification to the root-level pom.xml file resolves a dependency
conflict between rexster and hbase-server over grizzly2's exact
version number.