Add DistributedCache support in Faunus + Hadoop2
It's now possible to forego supplying a mapred job jar, in which case
Faunus under Hadoop 2 will iterate over all jars in the classpath and
call job.addFileToClassPath on each one, which puts them in the
DistributedCache and makes them available to the job at runtime.

This is only lightly tested.  It could also be generalized to Hadoop
1, although I'm not sure which types would be incompatible and which
can be shared and go in core.