Tweak lib file filter for zip/tar archives
The zipfile/tarfile assembly descriptor is now slightly more
inclusive.  It used to exclude any jar in the dependency tree with
classifier tests, javadoc, shared-resources, or example-data.  Of
those, tests could have potentially been problematic; there are some
dependencies in the titan-dist tree that in turn depend on artifacts
with classifier "tests" yet have dependency scope "compile".  Anything
with compile scope should go into lib, and now that's the case, even
if the classifier is "tests".  It's possible that these compile scopes
are just oversights in the poms of our dependencies, but I can't
easily tell and it's not worth the risk of excluding them.