Tweak titan-hadoop-core artifact deployment
The shared-resources and example-data artifacts are now detached from
the build, so they're no longer installed, signed, or deployed by
default.  I added explicit maven-install-plugin executions to install
them locally to support the titan-hadoop-1 and -2 modules.  The point
of this is to avoid deploying these artifacts to Maven Central.

This commit also redeclares the gpg plugin at the bottom of the
titan-hadoop-core pom so that the test jars will get signed.  Seems
like the maven-jar-plugin's test-jar execution is ordered after
maven-gpg-plugin's sign execution.