Giving up on titan-hadoop-core deploy tweaks
There is no way to exclude specific classifiers from the deploy stage
short of detaching the from the build.  Even disabling the
default-deploy execution and manually configuring deploy-file
executions for only a whitelist of classifiers is a deadend strategy,
because deploy-file requires an explicit repositoryId instead of just
defaulting to whatever repository is configured in the pom.  I forgot
how much of a headache this is, but the only way to get this behavior
is more modules.  Not going down that path right now...