Make Faunus copy DistributedCached'd jars to HDFS
Hadoop automagically copies files and creates symlinks when adding
jars to the DistributedCache under the local JobRunner, but Hadoop
requires the user (Faunus, in this case) to manually copy jars from
the local filesystem to HDFS when using an actual cluster.  This
commit extends AbstractDistCacheConfigurer and its subclasses to
address both cases instead of just working in the local case.