FaunusSerializer instantiation test
Creating a new FaunusSerializer for every edge processed by
EdgeCopyMapReduce ends up creating a separate Kryo for each of those
edges and consuming more than half of the CPU time under certain JVM
and workload configs.  This commit is an experimental change limited
to EdgeCopyMapReduce.java and is not a general fix -- just testing to
see if it alleviates the problem in testing.