Reduce num_tokens in zipfile cassandra.yaml
num_tokens=256, the upstream default, results in Faunus firing off at
least 256 mappers for every job, even if it's just Graph of the Gods.

Reduced num_tokens to 4 in the cassandra.yaml file that ships in the
Titan distribution archives, added a comment recommending increasing
it in production deployments.  This file is only used by titan.sh and
by the embedded Cassandra adapter right now.