Rename two addVertex overloads that take a label
Fix #702

Lead up to this commit:

TitanTransaction.addVertex(VertexLabel) and addVertex(String) both
take a vertex label argument.  TitanBlueprintsTransaction and TBGraph
define addVertex(Object).  The Object overload doesn't contain any
implementation logic by itself; it does instanceof checks on the
argument and calls another addVertex overload based on what it finds.
If the argument is a String, it calls the
TitanTransaction.addVertex(String vertexLabel) overload.

This behavior is incompatible with Blueprints.  Calling loadGraphML
(which in turn uses BP GraphMLReader and BatchGraph) on a file just
containing <node id="1"/> results in Blueprints calling
TitanTransaction.addVertex(Object) with the String "1".  Titan sees
that the argument is a String and attempts to interpret "1" as a
vertex label.  However, the implicit contract/intent on the Blueprints
side was to have "1" interpreted as a user-assigned ID (or, in the
specific case of Titan, to have Titan ignore that ID argument entirely
and self-assign an ID).

This commit renames the two TitanTransaction.addVertex methods that
take a vertex label to addVertexWithLabel.  It also modifies
TitanBlueprintsTransaction.addVertex(Object) to ignore arguments of
type String.  Now only addVertex arguments of type Number or
VertexLabel are considered by Titan.  Arguments of other types are
ignored.