Use shebang #!bin/bash for compat w/ #663
bash (4.2.37) refuses to parse the abs_path function when its invoked
as /bin/sh in the shebang line.  The three scripts in this commit die
with an error before doing anything useful because the shebang in
titan05 is /bin/sh.  Changing to /bin/bash lets it parse.
bin/cassandra was part of the original PR target (master) and also had
a bin/sh shebang on master.

elasticsearch uses /bin/sh and has its own recursive readlink logic
that apparently works fine in bash's sh compatibility mode.