Remove List-valued ConfigOption support
The only List-typed ConfigOption appears to be in
AbstractCassandraStoreManager.  That use can be served equally well by
a String[].  I converted it to ConfigOption<String[]> and removed
support for List-valued types in CommonsConfiguration.

List-valued types are problematic in the ConfigOption API since we
can't check the erased type of their generic parameter at runtime.