Option-ify Faunus's Kryo output buffer size
titan.hadoop.serializer.max-output-buffer-size now controls the
maxBufferSize parameter passed to com.esotericsoftware.kryo.io.Output
objects instantiated by Titan-Hadoop.  This is a hard upper-limit on
the size of serialized representations of objects written by
Titan-Hadoop.  It defaults to the same value which was hardcoded prior
to this commit, 10 MiB.