Rewrite ReflectiveConfigOptionLoader
This is now optimized for the case where only standard Titan config
options are in use (no third-party options), but falls back to a
slower implementation that can find third-party options if it fails to
resolve an option.

It's also got methods that support completely disabling it, in case
the user wants to takeover config option initialization, and for
configuring the classloader chain that it uses.  The default
classloader chain is

* ReflectiveConfigOptionLoader.class.getClassLoader()
* The current thread context class loader
* The classloader of the calling class

But this can be overridden through public methods at runtime.