doc CSS hack: make anchors offset by header height
Using intra-page anchor links with the documentation prior to this
commit was sort of clumsy, since the sticky/floating header at the top
of the viewport would obscure about the topmost ~85 pixels of whatever
the link jumped to -- usually the section header and an initial
sentence or two.

This CSS hack makes anchor links jump about 88 pixels too high to
compensate for the height of the floating/sticky header.