Test Cassandra SSL authentication
This commit adds subclasses of AstyanaxStoreTest and ThriftStoreTest
that run all the same test logic, except with SSL enabled in the store
manager configuration.  Since cassandra.yaml needs certain config
options set to enable SSL and point it at its keystore, this also
required changes to the test harness.  The pom got new executions of
the resources and surefire plugins; these filter SSLized
cassandra.yaml into target/cassandra/conf/localhost-murmur-ssl/ and
also copy (without resource filtering) a truststore and keystore to
the same directory.  The keystore is used by cassandra.yaml and the
truststore is used by the Titan configurations in AstyanaxSSLStoreTest
and ThriftSSLStore_Test.

The Astyanax test passes, but the Thrift test fails due to multiple
bugs in the CTConnectionFactory.makeRawConnection's SSL logic.  It
should be calling setTrustStore instead of setKeyStore, and it needs
to check whether the transport is already open before trying to open
it.