Check TSocket.isOpen before attempting .open
`CTConnectionFactory.makeRawConnection` calls `new TSocket(hostname,
port, timeout)` when SSL is not enabled.  This returns an unopened
socket which `makeRawConnection` opens before returning.  However,
when SSL is enabled, `makeRawConnection` forgoes the constructor in
favor of calling `TSSLTransportFactory.getClientSocket` to build an
SSLized TSocket.  The TSocket it returns is already opened, and
attempting to reopen it throws an exception.

This commit makes opening conditional on `!isOpen()`.

I found similar code in Hector a useful sanity check:

https://github.com/hector-client/hector/blob/hector-2.0-0/core/src/main/java/me/prettyprint/cassandra/connection/client/HThriftClient.java#L174-L176