Fix #740
I decided not to use the ConfigOption API.
GraphOfTheGodsFactory.create is likely to be seen/copied by Titan
users.  In the most recent discussion with Matthias about the
ConfigOption API, the verdict was to keep it internal to Titan insofar
as possible, so that we only have to manage stability on the string
config keys instead of both the string keys and ConfigOption
public-static-final-fields all over the code.