Tweaking timeouts in racy tests
Several index and TTL tests chronically fail because they're
intrinsically racy and rely on background threads/external processes
doing something or not doing something within a hard-coded timelimit.
This commit attempts to increase some of the most frequently failing
tests in CI and on my workstation, but this doesn't fix the problem.
This is not the long-term answer (relying on mocking/stubbing and
removing the sleeps to make index/TTL test logic completely
deterministic is the answer, and it will also make the tests much
faster).