Add $TITAN_JAVA_OPTS environment variable
Unlike the old $JAVA_OPTIONS, $TITAN_JAVA_OPTS does not overwrite the
default VM options with the value of the environment variable.
Keeping gremlin.sh's default logging-related VM arguments and
appending a few extras is almost always the desired behavior;
$JAVA_OPTIONS's overwrite behavior makes the common case cumbersome,
but I'm reluctant to delete or alter its semantics, since it's been in
gremlin.sh for ages and modifying it may break user scripts that
depend on the old behavior.  Hence the separate variable
$TITAN_JAVA_OPTS for the common case.