Tweaking titan-solr bdb & cass tests
The BDB- and Cassandra-backed tests of Solr introduced in this commit
don't pass.  There are still problems possibly related to
core/collection and field definition that fail the tests.  We may need
to modify or expand the schema files in titan-solr/src/test/resources,
or tweak the test code to use them correctly.

This commit also changes the default for SolrIndex.ZOOKEEPER_URL.  The
old default caused parse errors -- it seems like, despite the name and
default, the way the value is used in the code calls for a
comma-separated list of ZK quorum hosts.  Not sure on that change,
might have to revisit it.