Reimplement Faunus incremental loading (+ tests)
The config key for the incremental loading groovy script is currently
titan.hadoop.output.loader-script-file.  It should contain one or more
of the following methods:

* getOrCreateVertex(FaunusVertex, TitanGraph, Context, Logger)
* getOrCreateVertexProperty(TitanProperty, TitanVertex,
                            TitanGraph, Context, Logger)
* getOrCreateEdge(FaunusEdge, TitanVertex /* in */,
                  TitanVertex /* out */,
                  TitanGraph, Context, Logger)

The first method should return a TitanVertex and the last should
return a TitanEdge.  The return type of the middle method is currently
ignored.  In practice, the TitanProperty-typed parameter to the middle
method is always of the more specific FaunusProperty subtype, but
making this explicit would require a cast and doesn't seem to lend
additional functionality.  Might revisit this.

All of the above details are subject to change until this becomes part
of a release.