Update partitioned-vertex-label tests
This is a consequence of 628d48fe23431422e71225306ce2220611a45c86.
One of the test methods in TitanBaseGraphTest creates a partitioned
vertex label, but the ability to create these is now dependendent on
whether cluster.partition is true.  I made the label in
TitanBaseGraphTest non-partitioned moved the partitioned-vertex-label
related assertions to a smaller method in TitanPartitionGraphTest.