Kerberos+HBase+MapReduce authentication support
This commit adds support for HBase's authentication token system,
which is required for MapReduce (Titan-Hadoop/Faunus) jobs that want
to talk to a Kerberized HBase instance.

This commit has no effect on Kerberized HBase authentication outside
of MapReduce/Titan-Hadoop/Faunus, which works as described in the
HBase manual.

This commit was tested on CentOS 5, RHEL 6, and MIT Kerberos 1.8.x.
The Titan-Hadoop HBase read and write logic use different code and
config keys.  The read path uses HBase's own input format and config
keys, while the write path uses a custom format brought in with
Faunus.

When writing without HBase configuration files (including
hbase-site.xml) on the classpath of the MR tasks, the Titan-Hadoop
config setting
titan.hadoop.output.conf.ext.hbase.security.authentication should be
set to Kerberos.  This tells the HBaseStorageManager that starts up
inside the Titan-Hadoop write path to check for a Kerberos
authentication delegation token.  If an hbase-site.xml containing the
hbase.security.authentication=kerberos setting is on the MR task
classpath, then there's no need to duplicate it in the Titan-Hadoop
config file like this.

The read path just needs that setting on the submitting process's
classpath.